+++ __mock__/index.js
@@ -1,5 +1,6 @@
 import jsf from "json-schema-faker";
 import faker from "faker";
+import fs from "fs";
 
 const DATA_AMOUNT = 10;
 
@@ -28,3 +29,16 @@ export const createMockData = async () => {
     }
     return result;
 };
+
+createMockData().then(data => {
+    Object.keys(data).forEach(key => {
+        const jsonToWrite = {};
+        jsonToWrite[key] = data[key];
+        fs.writeFile(`__mock__/data/${key}.json`, JSON.stringify(jsonToWrite, null, 4), err => {
+            if (err) {
+                console.error(err);
+                throw new Error("Cannot write file");
+            }
+        });
+    });
+});
